refactor RecentProtocolRuns
