--- app/src/organisms/Desktop/Devices/HistoricalProtocolRun.module.css
@@ -1,5 +1,0 @@
-.protocol_name {                                                                                                   
-    overflow: hidden;                                                                                              
-    text-overflow: ellipsis;
-    white-space: nowrap;                                                                                             
-  }
