+++ app/src/organisms/Desktop/Devices/RecentProtocolRuns/HistoricalProtocolRunOverflowMenu.tsx
@@ -55,7 +55,7 @@ import {
 import { useIsRobotOnWrongVersionOfSoftware } from '/app/redux/robot-update'
 import { useIsEstopNotDisengaged } from '/app/resources/devices'
 
-import { useDownloadRunLog } from './hooks'
+import { useDownloadRunLog } from '../hooks'
 
 import type { MouseEventHandler } from 'react'
 import type { Run } from '@opentrons/api-client'
