+++ app/src/organisms/Desktop/Devices/RecentProtocolRuns/index.tsx
@@ -1,24 +1,7 @@
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import {
-  ALIGN_CENTER,
-  ALIGN_FLEX_START,
-  BasicButton,
-  BORDERS,
-  COLORS,
-  DIRECTION_COLUMN,
-  DISPLAY_FLEX,
-  Flex,
-  InfoScreen,
-  JUSTIFY_FLEX_END,
-  JUSTIFY_FLEX_START,
-  JUSTIFY_SPACE_BETWEEN,
-  LegacyStyledText,
-  SIZE_4,
-  SPACING,
-  StyledText,
-} from '@opentrons/components'
+import { BasicButton, InfoScreen, StyledText } from '@opentrons/components'
 import { useAllProtocolsQuery } from '@opentrons/react-api-client'
 
 import { useIsRobotViewable } from '/app/redux-resources/robots'
@@ -28,9 +11,9 @@ import {
   useRunStatuses,
 } from '/app/resources/runs'
 
-import { RECENT_PROTOCOL_RUNS_HEADER } from './constants'
-import { DeleteRecordsModal } from './DeleteRecordsModal'
+import { DeleteRecordsModal } from '../DeleteRecordsModal'
 import { HistoricalProtocolRun } from './HistoricalProtocolRun'
+import styles from './recentprotocolruns.module.css'
 
 interface RecentProtocolRunsProps {
   robotName: string
@@ -55,7 +38,10 @@ export function RecentProtocolRuns({
   }
 
   const robotIsBusy = currentRunId != null ? !isRunTerminal : false
+
+  // TODO (nd, 06/25/2026): audit this once full run delete endpoint is created
   const allRunsMutable = [...(runs ?? [])]
+
   return (
     <>
       {showDeleteRecordsModal ? (
@@ -67,31 +53,12 @@ export function RecentProtocolRuns({
           onConfirm={handleConfirmDeleteRuns}
         />
       ) : null}
-      <Flex
-        alignItems={ALIGN_FLEX_START}
-        backgroundColor={COLORS.white}
-        borderRadius={BORDERS.borderRadius8}
-        flexDirection={DIRECTION_COLUMN}
-        padding={`0 0 ${SPACING.spacing8}`}
-        width="100%"
-        marginBottom="6rem"
-      >
-        <Flex
-          padding={SPACING.spacing16}
-          borderBottom={BORDERS.lineBorder}
-          width="100%"
-          alignItems={ALIGN_CENTER}
-          gridGap={SPACING.spacing8}
-          justifyContent={JUSTIFY_SPACE_BETWEEN}
-        >
-          <StyledText desktopStyle="bodyLargeSemiBold" flex="1">
+      <div className={styles.container}>
+        <div className={styles.header}>
+          <StyledText desktopStyle="bodyLargeSemiBold">
             {t('run_history')}
           </StyledText>
-          <Flex
-            alignItems={ALIGN_CENTER}
-            justifyContent={JUSTIFY_FLEX_END}
-            gridGap={SPACING.spacing8}
-          >
+          <div className={styles.header_actions}>
             <BasicButton
               // TODO: wire up actions for downloading all
               onClick={() => {
@@ -108,27 +75,12 @@ export function RecentProtocolRuns({
             >
               {t('delete_all')}
             </BasicButton>
-          </Flex>
-        </Flex>
-        <Flex
-          alignItems={ALIGN_CENTER}
-          flexDirection={DIRECTION_COLUMN}
-          minHeight={SIZE_4}
-          padding={SPACING.spacing16}
-          width="100%"
-        >
+          </div>
+        </div>
+        <div className={styles.content}>
           {isRobotViewable && allRunsMutable && allRunsMutable?.length > 0 && (
             <>
-              <Flex
-                display="grid"
-                justifyContent={JUSTIFY_FLEX_START}
-                padding={SPACING.spacing8}
-                width="88%"
-                marginRight="12%"
-                gap={SPACING.spacing20}
-                color={COLORS.grey60}
-                gridTemplateColumns={RECENT_PROTOCOL_RUNS_HEADER}
-              >
+              <div className={styles.column_headers}>
                 <StyledText
                   desktopStyle="bodyDefaultRegular"
                   data-testid="RecentProtocolRuns_RunTitle"
@@ -159,19 +111,14 @@ export function RecentProtocolRuns({
                 >
                   {t('run_duration')}
                 </StyledText>
-              </Flex>
-              <Flex
-                flexDirection={DIRECTION_COLUMN}
-                gap={SPACING.spacing4}
-                width="100%"
-              >
+              </div>
+              <div className={styles.runs_list}>
                 {allRunsMutable
                   .sort(
                     (a, b) =>
                       new Date(b.createdAt).getTime() -
                       new Date(a.createdAt).getTime()
                   )
-
                   .map((run, index) => {
                     const protocol = protocols?.data?.data.find(
                       protocol => protocol.id === run.protocolId
@@ -193,25 +140,17 @@ export function RecentProtocolRuns({
                       />
                     )
                   })}
-              </Flex>
+              </div>
             </>
           )}
           {!isRobotViewable && (
             <InfoScreen content={t('offline_recent_protocol_runs')} />
           )}
           {isRobotViewable && allRunsMutable?.length === 0 && (
-            <LegacyStyledText
-              forwardedAs="p"
-              alignItems={ALIGN_CENTER}
-              display={DISPLAY_FLEX}
-              flex="1 0"
-              id="RecentProtocolRuns_no_runs"
-            >
-              {t('no_protocol_runs')}
-            </LegacyStyledText>
+            <InfoScreen content={t('no_protocol_runs')} />
           )}
-        </Flex>
-      </Flex>
+        </div>
+      </div>
     </>
   )
 }
